backend/pyproject.toml
[tool.poetry]
name = "educational-publishing-backend"
version = "0.1.0"
description = "Backend for Educational Publishing Platform"
authors = ["Admin <admin@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.1"}
pydantic = {extras = ["email"], version = "^2.4.2"}
pydantic-settings = "^2.0.3"
motor = "^3.3.1"
beanie = "^1.25.0"
python-multipart = "^0.0.9"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
redis = "^5.0.1"
boto3 = "^1.28.62"
httpx = "^0.27.2"
python-docx = "^1.1.0"
EbookLib = "^0.17.1"
celery = "^5.3.6"
mcp = "^1.20.0"
dnspython = "^2.4.0"
certifi = "^2024.0.0"


fastapi-sso = "^0.7.0"
google-api-python-client = "^2.100.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^1.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
